+++ reference-implementation/lib/UnderlyingSource.webidl
@@ -1,15 +1,15 @@
 dictionary UnderlyingSource {
-  ReadableStreamStartCallback start;
-  ReadableStreamPullCallback pull;
-  ReadableStreamCancelCallback cancel;
+  UnderlyingSourceStartCallback start;
+  UnderlyingSourcePullCallback pull;
+  UnderlyingSourceCancelCallback cancel;
   ReadableStreamType type;
   [EnforceRange] unsigned long long autoAllocateChunkSize;
 };
 
 typedef (ReadableStreamDefaultController or ReadableByteStreamController) ReadableStreamController;
 
-callback ReadableStreamStartCallback = any (ReadableStreamController controller);
-callback ReadableStreamPullCallback = Promise<void> (ReadableStreamController controller);
-callback ReadableStreamCancelCallback = Promise<void> (optional any reason);
+callback UnderlyingSourceStartCallback = any (ReadableStreamController controller);
+callback UnderlyingSourcePullCallback = Promise<void> (ReadableStreamController controller);
+callback UnderlyingSourceCancelCallback = Promise<void> (optional any reason);
 
- enum ReadableStreamType { "bytes" };
+enum ReadableStreamType { "bytes" };
