+++ reference-implementation/lib/abstract-ops/writable-streams.js
@@ -24,6 +24,7 @@ Object.assign(exports, {
   WritableStreamCloseQueuedOrInFlight,
   WritableStreamDefaultControllerClearAlgorithms,
   WritableStreamDefaultControllerError,
+  WritableStreamDefaultControllerErrorIfNeeded,
   WritableStreamDefaultWriterAbort,
   WritableStreamDefaultWriterClose,
   WritableStreamDefaultWriterCloseWithErrorPropagation,
