+++ reference-implementation/lib/index.js
@@ -2,9 +2,6 @@
 // This file is used as the entry point for browserifying the reference implementation to allow it
 // to run inside the wpt-runner "browser like" context.
 
-// const { TransformStream } = require('./transform-stream.js');
-
-// window.TransformStream = TransformStream;
 window.gc = gc;
 
 require('../generated/ByteLengthQueuingStrategy.js').install(window, ['Window']);
@@ -20,3 +17,6 @@ require('../generated/ReadableStreamBYOBRequest.js').install(window, ['Window'])
 require('../generated/WritableStream.js').install(window, ['Window']);
 require('../generated/WritableStreamDefaultWriter.js').install(window, ['Window']);
 require('../generated/WritableStreamDefaultController.js').install(window, ['Window']);
+
+require('../generated/TransformStream.js').install(window, ['Window']);
+require('../generated/TransformStreamDefaultController.js').install(window, ['Window']);
